+++ StMoviePlayer/StMoviePlayerGUI.h
@@ -1,5 +1,5 @@
 /**
- * Copyright © 2009-2023 Kirill Gavrilov <kirill@sview.ru>
+ * Copyright © 2009-2025 Kirill Gavrilov <kirill@sview.ru>
  *
  * StMoviePlayer program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -109,15 +109,22 @@ class StMoviePlayerGUI : public StGLRootWidget {
     ST_LOCAL virtual void stglDraw(unsigned int theView) ST_ATTR_OVERRIDE;
     ST_LOCAL virtual void stglUpdate(const StPointD_t& thePointZo,
                                      bool theIsPreciseInput) ST_ATTR_OVERRIDE;
+    ST_LOCAL virtual bool tryUnClick(const StClickEvent& theEvent,
+                                     bool& theIsItemUnclicked) ST_ATTR_OVERRIDE;
+    using StGLRootWidget::tryUnClick;
 
-    /**
-     * Handle gesture.
-     */
+        public: //! @name visibility
+
+    /** Track touches. */
+    ST_LOCAL void doTouch(const StTouchEvent& theEvent);
+
+    /** Handle gesture. */
     ST_LOCAL void doGesture(const StGestureEvent& theEvent);
-    ST_LOCAL bool isVisibleGUI() const { return myVisLerp.getValue() > 0.0; }
+    ST_LOCAL bool isVisibleGUI() const { return !myAnimVisibility.isHidden(); }
     ST_LOCAL void setVisibility(const StPointD_t& theCursor,
                                 bool theToForceHide = false,
                                 bool theToForceShow = false);
+    ST_LOCAL void updateDescLabel(const StPointD_t& theCursor, StFormat theSrcFormat, StViewSurface theViewMode);
 
         public:
 
@@ -217,10 +224,7 @@ class StMoviePlayerGUI : public StGLRootWidget {
     StMoviePlayer*      myPlugin;           //!< link to the main Drawer class
     StWindow*           myWindow;           //!< link to the window instance
     StTranslations*     myLangMap;          //!< translated strings map
-    StTimer             myVisibilityTimer;  //!< minimum visible delay
-    StTimer             myEmptyTimer;       //!< empty list delay
-    StTimer             myTapTimer;         //!< single tap delay
-    StGLAnimationLerp   myVisLerp;
+    StGLAnimVisibility  myAnimVisibility;
 
     StGLImageRegion*    myImage;            //!< the main video frame
     StGLSubtitles*      mySubtitles1;       //!< the subtitles
@@ -270,7 +274,6 @@ class StMoviePlayerGUI : public StGLRootWidget {
 
     StGLTable*          myHKeysTable;
 
-    bool                myIsVisibleGUI;
     bool                myIsExperimental;
     int                 myIconStep;         //!< step between icons, in pixels
     int                 myBottomBarNbLeft;  //!< number of icons at left  side of bottom bar, in pixels