+++ CHANGELOG.md
@@ -14,6 +14,8 @@ sView ??.?? (in development)
 - Linux, enable building with `libopenvr-dev` for **OpenVR** support (untested).
 - Windows, removed 32-bit binaries from package.
 - Windows, removed FreeImage and DevIL libraries from package.
+- Android, taps on touchsreen now hide/show user interface.
+- Android, OpenAL soft has been upgraded to 1.24.2.
 - macOS, DMG packages are now built for ARM64.
 - Removed legacy **NPAPI**/**ActiveX** plugin (browsers dropped support for native plugins since 2015-2018).
 - Configuration, building environment moved to *CMake*.